+++ prisma/migrations/20260904165554_freelancer_recorrencia/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "payment_requests" ADD COLUMN     "recurrent" BOOLEAN NOT NULL DEFAULT false,
+ADD COLUMN     "weekCount" INTEGER;